+++ examples/polkadot/push_wasm_client_code_test.go
@@ -22,7 +22,7 @@ import (
 )
 
 const (
-	heightDelta      = int64(20)
+	heightDelta      = 20
 	votingPeriod     = "30s"
 	maxDepositPeriod = "10s"
 )
